requirements.txt
fastapi
uvicorn[standard]
requests
gspread
python-dotenv
